move testOracle files to test.shared folder
