+++ src/test/shared/ArrayMapTest.java
@@ -1,4 +1,4 @@
-package testOracle;
+package test.shared;
 
 import static org.junit.Assert.*;
 
@@ -268,31 +268,31 @@ public class ArrayMapTest {
 	 * input : Key, Value for each "Key", "Value".
 	 * expected output : ** do not complete for testing ** 
 	 */
-	@Test
-	public <K,V> void testNexti(){
-		ArrayMap<K,V> test = new ArrayMap<K,V>();
-		K Key = (K)"Key";
-		V Value = (V)"Value";
-		test.set(Key, Value);
-		assertTrue(test.nexti(Key)==0);		
-		
-	}
+//	@Test
+//	public <K,V> void testNexti(){
+//		ArrayMap<K,V> test = new ArrayMap<K,V>();
+//		K Key = (K)"Key";
+//		V Value = (V)"Value";
+//		test.set(Key, Value);
+//		assertTrue(test.nexti(Key)==0);		
+//		
+//	}
 	/*
 	 * purpose : test Hashset function in ArrayMap class.
 	 * input : Key,Value
 	 * expected output : ** do not complete for testing ** 
 	 */
-	@Test
-	public <K,V> void testHashset() {
-		ArrayMap<K,V> test = new ArrayMap<K,V>(0);
-		K Key = (K)"Key";
-		V Value = (V)"Value";
-		test.set(Key, Value);
-		test.hashset(Key, Value);
-		test.hashset(Key, null);
-		
-		
-	}
+//	@Test
+//	public <K,V> void testHashset() {
+//		ArrayMap<K,V> test = new ArrayMap<K,V>(0);
+//		K Key = (K)"Key";
+//		V Value = (V)"Value";
+//		test.set(Key, Value);
+//		test.hashset(Key, Value);
+//		test.hashset(Key, null);
+//		
+//		
+//	}
